+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add lifecycle `processingInterval` setting support for `processing_interval`.
+
 ## 1.21.0-beta.0 - 2026-07-14
 
 ### Added